authentication and privilegies

+++ blog-bloks/app/Http/Controllers/PostController.php
@@ -24,8 +24,7 @@ class PostController extends Controller
     public function store(PostFormRequest $request)
     {
         $validated = $request->validated();
-
-        $post = Post::create($validated); 
+        $post = $request->user()->posts()->create($validated);
 
         return redirect()
         ->route('posts.show', [$post])
@@ -49,6 +48,7 @@ class PostController extends Controller
      */
     public function edit(Post $post)
     {
+        $this->authorize('update', $post);
         return view('posts.edit', [
             'post' => $post,
         ]);
@@ -59,6 +59,8 @@ class PostController extends Controller
      */
     public function update(PostFormRequest $request, Post $post)
     {
+        $this->authorize('update', $post);
+
         $validated = $request->validated();
 
         $post->update($validated);
@@ -73,6 +75,8 @@ class PostController extends Controller
      */
     public function destroy(Post $post)
     {
+        $this->authorize('delet', $post);
+        
         $post->delete();
 
         return redirect()
